+++ Swift Vocabulary/Swift Vocabulary/DefinitionViewController.swift
@@ -10,13 +10,22 @@ import UIKit
 
 class DefinitionViewController: UIViewController {
 
+    var vocabWord: VocabularyWord?
+    
     override func viewDidLoad() {
         super.viewDidLoad()
 
         // Do any additional setup after loading the view.
+        
+        updateViews()
     }
     
 
+    func updateViews() {
+        if let unwrappedVocabWord = vocabWord {
+             = unwrappedVocabWord.word
+        }
+    }
     /*
     // MARK: - Navigation
 
@@ -26,5 +35,4 @@ class DefinitionViewController: UIViewController {
         // Pass the selected object to the new view controller.
     }
     */
-
 }
