--- Swift Vocabulary/Swift Vocabulary/worldsTableViewController.swift
@@ -1,37 +1,0 @@
-//
-//  worldsTableViewController.swift
-//  Swift Vocabulary
-//
-//  Created by Ray Wellman on 2/17/20.
-//  Copyright © 2020 Richard Ray Wellman. All rights reserved.
-//
-
-import UIKit
-
-class worldsTableViewController: UITableViewController {
-    
-    var vocabWords: [VocabularyWord] = [VocabularyWord(word: "", definition: "")]
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        // Uncomment the following line to preserve selection between presentations
-        self.clearsSelectionOnViewWillAppear = false
-
-        // Uncomment the following line to display an Edit button in the navigation bar for this view controller.
-        self.navigationItem.rightBarButtonItem = self.editButtonItem
-    }
-
-    // MARK: - Table view data source
-
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        // #warning Incomplete implementation, return the number of rows
-        return 1
-    }
-    
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-}
